Provide programmatic endpoint, don't require test to depend on shadowJar
