+++ jfr-streaming/src/main/java/io/opentelemetry/contrib/jfr/metrics/Constants.java
@@ -7,7 +7,7 @@ package io.opentelemetry.contrib.jfr.metrics;
 
 import io.opentelemetry.api.common.AttributeKey;
 
-public class Constants {
+public final class Constants {
   private Constants() {}
 
   public static final String KILOBYTES = "KB";
