+++ jfr-streaming/src/test/java/io/opentelemetry/contrib/jfr/metrics/AbstractMetricsTest.java
@@ -8,6 +8,7 @@ package io.opentelemetry.contrib.jfr.metrics;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 
+import io.opentelemetry.contrib.jfr.JfrMetrics;
 import io.opentelemetry.sdk.metrics.SdkMeterProvider;
 import io.opentelemetry.sdk.metrics.data.MetricData;
 import io.opentelemetry.sdk.metrics.testing.InMemoryMetricReader;
@@ -27,6 +28,7 @@ public class AbstractMetricsTest {
     metricReader = new InMemoryMetricReader();
     meterProvider =
         SdkMeterProvider.builder().registerMetricReader(metricReader).buildAndRegisterGlobal();
+    JfrMetrics.enableJfr(meterProvider);
   }
 
   protected void waitAndAssertMetrics(Consumer<MetricDataAssert>... assertions) {
